+++ src/main/java/beyond/samdasoo/lead/dto/LeadResponseDto.java
@@ -14,9 +14,9 @@ public class LeadResponseDto {
     private Long leadNo;       // 리드 번호
     private String name;       // 리드 이름
     private LeadStatus status; // 리드 상태
-    private int expSales;      // 예상 매출
+    private Long expSales;      // 예상 매출
     private int expMargin;     // 예상 이익률
-    private int expProfit;     // 예상 이익금
+    private Long expProfit;     // 예상 이익금
     private Long process;       // 프로세스
     private Long subProcess;    // 서브 프로세스
     private int successPer;    // 성공 확률
